% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxed_radio_subgroup.R
\name{boxed_radio_subgroup}
\alias{boxed_radio_subgroup}
\title{UI Function: grouped radio button within a Shinydashboard box}
\usage{
boxed_radio_subgroup(input_id, id, box_title = "", box_status = "primary",
  label = "", choices = NULL, inline = FALSE, box_width = NULL)
}
\arguments{
\item{input_id}{The input slot that will be used to access the value.}

\item{id}{The id for button list (should be unique within the group).}

\item{box_title}{Title for the container box (optional).}

\item{box_status}{The status of the item This determines the item's background color (see \code{\link[shinydashboard]{box}}).}

\item{label}{Display label for the control, or NULL for no label.}

\item{choices}{List of values to select from (if elements of the list are named then that name rather than the value is displayed to the user).}

\item{inline}{If TRUE, render the choices inline (i.e. horizontally).}

\item{box_width}{The width of the box, using the Bootstrap grid system.}
}
\description{
This is a radio button UI element that can be linked to others.
}
\note{
This is not a shiny module.
}
