% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_app.R
\name{launch_app}
\alias{launch_app}
\title{Run the Shiny App}
\usage{
launch_app(pkg_name)
}
\arguments{
\item{pkg_name}{The name of the package.}
}
\description{
This function runs \code{\link[shiny]{runApp}} with \code{./inst/application} set as the target directory.
}
