% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_about.R
\name{mod_about}
\alias{mod_about}
\alias{tabItemContentUI_about}
\alias{tabItemContent_about}
\title{Shiny Module: About text in a shinydashboard box.}
\usage{
tabItemContentUI_about(id, tab_name)

tabItemContent_about(input, output, session)
}
\arguments{
\item{id}{Output id}

\item{tab_name}{Name of the tab panel}

\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}
}
\description{
About text in a shinydashboard box.
}
