% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_documents.R
\name{mod_documents}
\alias{mod_documents}
\alias{tabItemContentUI_documents}
\alias{tabItemContent_documents}
\title{Shiny Module: Documents}
\usage{
tabItemContentUI_documents(id, tab_name)

tabItemContent_documents(input, output, session)
}
\arguments{
\item{id}{Output id}

\item{tab_name}{Name of the tab panel}

\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}
}
\description{
Documents
}
