% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_filters.R
\name{mod_filters}
\alias{mod_filters}
\alias{condUI_filters}
\alias{cond_filters}
\title{Shiny Module: Filter UI linked to Menu Items.}
\usage{
condUI_filters(id, tab_names)

cond_filters(input, output, session)
}
\arguments{
\item{id}{Output id}

\item{tab_names}{The names of the Menu Items that should trigger this conditional panel.}

\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}
}
\description{
Filter UI linked to Menu Items (using conditionalPanel).
}
