% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_lists.R
\name{mod_lists}
\alias{mod_lists}
\alias{tabItemContentUI_lists}
\alias{tabItemContent_lists}
\title{Shiny Module:_grouped radio button within a Shinydashboard box}
\usage{
tabItemContentUI_lists(id, tab_name)

tabItemContent_lists(input, output, session)
}
\arguments{
\item{id}{Output id}

\item{tab_name}{Name of the tab panel}

\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}
}
\description{
This is a radio button UI element that can be linked to others.
}
