% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_map.R
\name{mod_map}
\alias{mod_map}
\alias{tabItemContentUI_map}
\alias{tabItemContent_map}
\title{Shiny Module: fullscreen leaflet map in a tab panel}
\usage{
tabItemContentUI_map(id, tab_name)

tabItemContent_map(input, output, session)
}
\arguments{
\item{id}{Output id}

\item{tab_name}{Name of the tab panel}

\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}
}
\description{
A fullscreen leaflet map in a tab panel.
}
