% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_projectCard.R
\name{mod_projectCard}
\alias{mod_projectCard}
\alias{projectCardUI}
\alias{projectCard}
\title{Shiny Module: Project Card}
\usage{
projectCardUI(id)

projectCard(input, output, session)
}
\arguments{
\item{id}{Output id}

\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}
}
\description{
Project Card
}
