% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_table.R
\name{mod_table}
\alias{mod_table}
\alias{tabItemContentUI_table}
\alias{tabItemContent_table}
\title{Shiny Module: Table}
\usage{
tabItemContentUI_table(id, tab_name)

tabItemContent_table(input, output, session)
}
\arguments{
\item{id}{Output id}

\item{tab_name}{Name of the tab panel}

\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}
}
\description{
Table
}
