% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_template.R
\name{module_name}
\alias{module_name}
\alias{ui_function_name}
\alias{server_function_name}
\title{Shiny Module: _description here_}
\usage{
ui_function_name(id, tab_name)

server_function_name(input, output, session)
}
\arguments{
\item{id}{Output id}

\item{tab_name}{Name of the tab panel}

\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}
}
\description{
This is the UI portion of this module.
}
