% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_obj.R
\name{server_function_name}
\alias{server_function_name}
\title{Shiny Module: _description here_}
\usage{
server_function_name(input, output, session)
}
\arguments{
\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}
}
\description{
This is the server portion of this module.
}
\seealso{
\code{\link{tabItemContentUI_map}}
}
