% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sidebar_bottom.R
\name{sidebar_bottom}
\alias{sidebar_bottom}
\title{CSS Styling for UI in the Lower Sidebar of a Shinydashboard.}
\arguments{
\item{...}{UI content.}
}
\description{
CSS Styling for UI in the Lower Sidebar of a Shinydashboard.
}
