% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabItemContent_about.R
\name{tabItemContent_about}
\alias{tabItemContent_about}
\title{Shiny Module: About text in a shinydashboard box.}
\usage{
tabItemContent_about(input, output, session)
}
\arguments{
\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}
}
\description{
This is the server portion of this module.
}
\seealso{
\code{\link{tabItemContentUI_about}}
}
