% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabItemContent_map.R
\name{tabItemContent_map}
\alias{tabItemContent_map}
\title{Shiny Module: fullscreen leaflet map in a tab panel}
\usage{
tabItemContent_map(input, output, session)
}
\arguments{
\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}
}
\description{
This is the server portion of this module.
}
\seealso{
\code{\link{tabItemContentUI_map}}
}
