% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabItemContentUI_about.R
\name{tabItemContentUI_about}
\alias{tabItemContentUI_about}
\title{Shiny Module: About text in a shinydashboard box.}
\usage{
tabItemContentUI_about(id, tab_name)
}
\arguments{
\item{id}{Output id}

\item{tab_name}{Name of the tab panel}
}
\description{
This is the UI portion of this module.
}
\seealso{
\code{\link{tabItemContent_about}}
}
