% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabItemContentUI_map.R
\name{tabItemContentUI_map}
\alias{tabItemContentUI_map}
\title{Shiny Module: fullscreen leaflet map in a tab panel}
\usage{
tabItemContentUI_map(id, tab_name)
}
\arguments{
\item{id}{Output id}

\item{tab_name}{Name of the tab panel}
}
\description{
This is the UI portion of this module.
}
\seealso{
\code{\link{tabItemContent_map}}
}
