% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templateUI_obj.R
\name{ui_function_name}
\alias{ui_function_name}
\title{Shiny Module: _description here_}
\usage{
ui_function_name(id, tab_name)
}
\arguments{
\item{id}{Output id}

\item{tab_name}{Name of the tab panel}
}
\description{
This is the UI portion of this module.
}
\seealso{
\code{\link{tabItemContent_map}}
}
