% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_projectCard.R
\name{mod_projectCard}
\alias{mod_projectCard}
\alias{tabItemContentUI_projectCard}
\alias{tabItemContent_projectCard}
\title{Shiny Module: Project Card}
\usage{
tabItemContentUI_projectCard(id)

tabItemContent_projectCard(input, output, session)
}
\arguments{
\item{id}{Output id}

\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}

\item{tab_name}{Name of the tab panel}
}
\description{
Project Card
}
